==> Scripts/09_tables.R <==
################################################################################
######################### Table DR - IC50 MTT GSH1 #############################
################################################################################

library(stargazer)

# Criando o dataframe
data <- data.frame(
  População = c("Referência", "POP - GSH1", "C7 - GSH1", "C6 - GSH1", "C44 - GSH1", 
                "C58 - GSH1", "C73 - GSH1", "C85 - GSH1", "C76 - PGPA",
                "C67p - PGPA", "C68 - PGPA"),
  ED50 = c(144.624, 150.905, "87.048*", "96.104*", "63.572*", 141.772, 145.682, 97.882,
           "27.539*", "39.210*", "63.445*"),
  `Erro padrão` = c(72.465, 76.088, 20.664, 24.714, 10.264, 114.900, 92.312, 
                    32.522, 7.618, 11.147, 10.925),
  `P-valor` = c(0.051901,
                0.059955,
                0.000359,
                0.000791,
                0.000003,
                0.230268,
                0.128801,
                0.006672,
                0.001536,
                0.001940,
                0.000008),
  `Intervalo de confiança` = c("(-1.24 - 290.49)", "(-6.89 - 308.70)", 
                               "(44.19 - 129.90)", "(44.85 - 147.36)", 
                               "(42.28 - 84.86)", "(-96.52 - 380.06)", 
                               "(-45.76 - 337.12)", "(30.25 - 165.52)",
                               "(11.74 - 43.34)", "(16.09 - 62.33)",
                               "(40.79 - 86.10)"))

# Exibindo o dataframe
print(data)

stargazer(data, summary = F)
